refactor: 이름 수정 및 mailTarget 메소드 서비스 위치 변경
