+++ src/main/kotlin/apply/application/MailTargetService.kt
@@ -4,6 +4,7 @@ import apply.domain.evaluationtarget.EvaluationStatus
 import apply.domain.evaluationtarget.EvaluationTarget
 import apply.domain.evaluationtarget.EvaluationTargetRepository
 import apply.domain.user.UserRepository
+import apply.domain.user.findAllByEmailIn
 import org.springframework.stereotype.Service
 import javax.transaction.Transactional
 
@@ -17,6 +18,12 @@ class MailTargetService(
         val userIds = findEvaluationTargets(evaluationId, evaluationStatus).map { it.userId }
         return userRepository.findAllById(userIds)
             .map { MailTargetResponse(it.email, it.name) }
+    }
+
+    fun findAllByEmails(emails: List<String>): List<MailTargetResponse> {
+        val users = userRepository.findAllByEmailIn(emails)
+        val anonymousEmails = emails - users.map { it.email }
+        return users.map { MailTargetResponse(it) } + anonymousEmails.map { MailTargetResponse(it) }
     }
 
     private fun findEvaluationTargets(evaluationId: Long, evaluationStatus: EvaluationStatus?): List<EvaluationTarget> {
