+++ src/test/kotlin/apply/application/MailTargetServiceTest.kt
@@ -11,6 +11,7 @@ import apply.domain.evaluationtarget.EvaluationStatus.PENDING
 import apply.domain.evaluationtarget.EvaluationStatus.WAITING
 import apply.domain.evaluationtarget.EvaluationTargetRepository
 import apply.domain.user.UserRepository
+import apply.domain.user.findAllByEmailIn
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.verify
@@ -108,5 +109,22 @@ class MailTargetServiceTest {
         val actual = mailTargetService.findMailTargets(EVALUATION_ID, WAITING)
         assertThat(actual).hasSize(1)
         assertThat(actual[0].email).isEqualTo("waiting@email.com")
+    }
+
+    @Test
+    fun `이메일 중에서 회원에 해당하는 이메일이 있으면 (회원이름, 이메일)을, 회원에 해당하는 이메일이 없으면 (공백, 이메일)을 반환한다`() {
+        val users = listOf(
+            createUser(name = "회원1", email = "test1@email.com"),
+            createUser(name = "회원3", email = "test3@email.com")
+        )
+        val emails = listOf("test1@email.com", "test2@email.com", "test3@email.com")
+        val mailTargetResponses = listOf(
+            MailTargetResponse("test1@email.com", "회원1"),
+            MailTargetResponse("test3@email.com", "회원3"),
+            MailTargetResponse("test2@email.com")
+        )
+        every { userRepository.findAllByEmailIn(emails) } returns users
+        val actual = mailTargetService.findAllByEmails(emails)
+        assertThat(actual).isEqualTo(mailTargetResponses)
     }
 }
